backend/requirements.txt
# Django and Django REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0

# Channels for WebSockets
channels==4.0.0
daphne==4.0.0

# Database
dj-database-url==2.1.0

# Environment variables
python-dotenv==1.0.0

# Cryptography
cryptography==41.0.4

# File handling
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# API documentation
drf-yasg==1.21.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utils
python-dateutil==2.8.2

# Optional dependencies for production
# Uncomment these for production deployment
# psycopg2-binary==2.9.9  # For PostgreSQL
# channels-redis==4.1.0   # For Redis channel layer 